requirements.txt
appium-python-client
pyserial
PyYAML
robotframework
selenium
webdriver-manager
packaging
robotframework-httpctrl